refactor: extract business logic from controllers into smaller service layers
