+++ API/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "nodemon --exec ts-node src/index.ts",
-    "seed": "ts-node src/seed/create-seed.ts"
+    "seed": "ts-node src/seed/createSeed.ts"
   },
   "author": "",
   "license": "ISC",
